+++ app/src/main/java/woowacourse/shopping/model/order/discount/DiscountPolicy.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.model.order.discount
+
+import woowacourse.shopping.model.order.OrderItem
+import woowacourse.shopping.model.product.Money
+
+interface DiscountPolicy {
+    fun calculateDiscount(
+        items: List<OrderItem>,
+        totalProductAmount: Money,
+        shippingFee: Money,
+    ): Money
+}
