+++ app/src/main/java/woowacourse/shopping/model/order/discount/FreeShippingDiscountPolicy.kt
@@ -1,0 +1,17 @@
+package woowacourse.shopping.model.order.discount
+
+import woowacourse.shopping.model.order.OrderItem
+import woowacourse.shopping.model.product.Money
+
+class FreeShippingDiscountPolicy(
+    private val minimumAmount: Money,
+) : DiscountPolicy {
+    override fun calculateDiscount(
+        items: List<OrderItem>,
+        totalProductAmount: Money,
+        shippingFee: Money,
+    ): Money {
+        if (totalProductAmount < minimumAmount) return Money.ZERO
+        return shippingFee
+    }
+}
