+++ app/src/main/java/woowacourse/shopping/model/order/shipping/ShippingPolicy.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.model.order.shipping
+
+import woowacourse.shopping.model.product.Money
+
+interface ShippingPolicy {
+    fun calculateShippingFee(totalProductAmount: Money): Money
+}
